Made a method to resolve this issue. https://github.com/Steven-Chorkawy/Submit-AR-Invoice/issues/96#issuecomment-738789131
